+++ .gitignore
@@ -1,15 +1,16 @@
 *.iml
 .gradle
 /local.properties
-/.idea/caches/build_file_checksums.ser
+/.idea/caches
 /.idea/libraries
 /.idea/modules.xml
 /.idea/workspace.xml
+/.idea/navEditor.xml
+/.idea/assetWizardSettings.xml
 .DS_Store
 /build
 /captures
 .externalNativeBuild
-
 
 # Created by https://www.gitignore.io/api/java,gradle,android,androidstudio
 # Edit at https://www.gitignore.io/?templates=java,gradle,android,androidstudio
