+++ app/src/main/java/com/comercial/iruber/cliente/dao/ClienteDAO.java
@@ -1,0 +1,4 @@
+package com.comercial.iruber.cliente.dao;
+
+public class ClienteDAO {
+}
