+++ app/src/main/java/com/comercial/iruber/cliente/dominio/Cartao.java
@@ -1,0 +1,51 @@
+package com.comercial.iruber.cliente.dominio;
+
+import java.util.Date;
+
+public class Cartao {
+    private long idCartao;
+    private String numero;
+    private String nome;
+    private int codSeguranca;
+    private Date dataVencimento;
+
+    public long getIdCartao() {
+        return idCartao;
+    }
+
+    public void setIdCartao(long idCartao) {
+        this.idCartao = idCartao;
+    }
+
+    public String getNumero() {
+        return numero;
+    }
+
+    public void setNumero(String numero) {
+        this.numero = numero;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public int getCodSeguranca() {
+        return codSeguranca;
+    }
+
+    public void setCodSeguranca(int codSeguranca) {
+        this.codSeguranca = codSeguranca;
+    }
+
+    public Date getDataVencimento() {
+        return dataVencimento;
+    }
+
+    public void setDataVencimento(Date dataVencimento) {
+        this.dataVencimento = dataVencimento;
+    }
+}
