+++ app/src/main/java/com/comercial/iruber/cliente/dominio/Pessoa.java
@@ -1,0 +1,40 @@
+package com.comercial.iruber.cliente.dominio;
+
+public class Pessoa {
+    private long idPessoa;
+    private String nome;
+    private String cpf;
+    private String email;
+
+    public long getIdPessoa() {
+        return idPessoa;
+    }
+
+    public void setIdPessoa(long idPessoa) {
+        this.idPessoa = idPessoa;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public String getCpf() {
+        return cpf;
+    }
+
+    public void setCpf(String cpf) {
+        this.cpf = cpf;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+}
