+++ app/src/main/java/com/comercial/iruber/pedido/dominio/ItemPedido.java
@@ -1,0 +1,34 @@
+package com.comercial.iruber.pedido.dominio;
+
+import java.math.BigDecimal;
+
+public class ItemPedido {
+    private long idItemPedido;
+    private BigDecimal valor;
+    private int quantidade;
+
+
+    public long getIdItemPedido() {
+        return idItemPedido;
+    }
+
+    public void setIdItemPedido(long idItemPedido) {
+        this.idItemPedido = idItemPedido;
+    }
+
+    public BigDecimal getValor() {
+        return valor;
+    }
+
+    public void setValor(BigDecimal valor) {
+        this.valor = valor;
+    }
+
+    public int getQuantidade() {
+        return quantidade;
+    }
+
+    public void setQuantidade(int quantidade) {
+        this.quantidade = quantidade;
+    }
+}
