+++ app/src/main/java/com/comercial/iruber/restaurante/dao/RestauranteDAO.java
@@ -1,0 +1,4 @@
+package com.comercial.iruber.restaurante.dao;
+
+public class RestauranteDAO {
+}
