+++ app/src/main/java/com/comercial/iruber/restaurante/dominio/Ingrediente.java
@@ -1,0 +1,31 @@
+package com.comercial.iruber.restaurante.dominio;
+
+public class Ingrediente {
+    private long idIngrediente;
+    private String nome;
+    private boolean disponivel;
+
+    public long getIdIngrediente() {
+        return idIngrediente;
+    }
+
+    public void setIdIngrediente(long idIngrediente) {
+        this.idIngrediente = idIngrediente;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public boolean isDisponivel() {
+        return disponivel;
+    }
+
+    public void setDisponivel(boolean disponivel) {
+        this.disponivel = disponivel;
+    }
+}
