+++ app/src/main/java/com/comercial/iruber/restaurante/dominio/Prato.java
@@ -1,0 +1,61 @@
+package com.comercial.iruber.restaurante.dominio;
+
+import java.math.BigDecimal;
+import java.util.ArrayList;
+
+public class Prato {
+    private long idProduto;
+    private String nome;
+    private String descricao;
+    private boolean disponivel;
+    private BigDecimal valor;
+    private ArrayList<Ingrediente> ingredientes;
+
+    public long getIdProduto() {
+        return idProduto;
+    }
+
+    public void setIdProduto(long idProduto) {
+        this.idProduto = idProduto;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public String getDescricao() {
+        return descricao;
+    }
+
+    public void setDescricao(String descricao) {
+        this.descricao = descricao;
+    }
+
+    public boolean isDisponivel() {
+        return disponivel;
+    }
+
+    public void setDisponivel(boolean disponivel) {
+        this.disponivel = disponivel;
+    }
+
+    public BigDecimal getValor() {
+        return valor;
+    }
+
+    public void setValor(BigDecimal valor) {
+        this.valor = valor;
+    }
+
+    public ArrayList<Ingrediente> getIngredientes() {
+        return ingredientes;
+    }
+
+    public void setIngredientes(ArrayList<Ingrediente> ingredientes) {
+        this.ingredientes = ingredientes;
+    }
+}
