+++ app/src/main/java/com/comercial/iruber/usuario/dominio/Cidade.java
@@ -1,0 +1,22 @@
+package com.comercial.iruber.usuario.dominio;
+
+public class Cidade {
+    private long idCidade;
+    private String nome;
+
+    public long getIdCidade() {
+        return idCidade;
+    }
+
+    public void setIdCidade(long idCidade) {
+        this.idCidade = idCidade;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+}
