+++ app/src/main/java/com/comercial/iruber/usuario/dominio/Endereco.java
@@ -1,0 +1,40 @@
+package com.comercial.iruber.usuario.dominio;
+
+public class Endereco {
+    private long idEndereco;
+    private String logradouro;
+    private int numero;
+    private Cidade cidade;
+
+    public long getIdEndereco() {
+        return idEndereco;
+    }
+
+    public void setIdEndereco(long idEndereco) {
+        this.idEndereco = idEndereco;
+    }
+
+    public String getLogradouro() {
+        return logradouro;
+    }
+
+    public void setLogradouro(String logradouro) {
+        this.logradouro = logradouro;
+    }
+
+    public int getNumero() {
+        return numero;
+    }
+
+    public void setNumero(int numero) {
+        this.numero = numero;
+    }
+
+    public Cidade getCidade() {
+        return cidade;
+    }
+
+    public void setCidade(Cidade cidade) {
+        this.cidade = cidade;
+    }
+}
