--- app/src/main/java/com/example/misael/iruber/MainActivity.java
@@ -1,13 +1,0 @@
-package com.example.misael.iruber;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
