--- app/src/main/java/dominio/endereco/Cidade.java
@@ -1,22 +1,0 @@
-package dominio.endereco;
-
-public class Cidade {
-    private long idCidade;
-    private String nome;
-
-    public long getIdCidade() {
-        return idCidade;
-    }
-
-    public void setIdCidade(long idCidade) {
-        this.idCidade = idCidade;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-}
