--- app/src/main/java/dominio/pedido/ItemPedido.java
@@ -1,34 +1,0 @@
-package dominio.pedido;
-
-import java.math.BigDecimal;
-
-public class ItemPedido {
-    private long idItemPedido;
-    private BigDecimal valor;
-    private int quantidade;
-
-
-    public long getIdItemPedido() {
-        return idItemPedido;
-    }
-
-    public void setIdItemPedido(long idItemPedido) {
-        this.idItemPedido = idItemPedido;
-    }
-
-    public BigDecimal getValor() {
-        return valor;
-    }
-
-    public void setValor(BigDecimal valor) {
-        this.valor = valor;
-    }
-
-    public int getQuantidade() {
-        return quantidade;
-    }
-
-    public void setQuantidade(int quantidade) {
-        this.quantidade = quantidade;
-    }
-}
