--- app/src/main/java/dominio/pedido/Pedido.java
@@ -1,84 +1,0 @@
-package dominio.pedido;
-
-import java.math.BigDecimal;
-import java.util.ArrayList;
-import java.util.Date;
-
-import dominio.cliente.Cliente;
-import dominio.restaurante.Entregador;
-import dominio.restaurante.Restaurante;
-
-public class Pedido {
-    private long idPedido;
-    private Cliente cliente;
-    private Entregador entregador;
-    private Restaurante restaurante;
-    private ArrayList<ItemPedido> itensPedido;
-    private Date data;
-    private BigDecimal valorTotal;
-    private StatusPedido statusPedido;
-
-    public long getIdPedido() {
-        return idPedido;
-    }
-
-    public void setIdPedido(long idPedido) {
-        this.idPedido = idPedido;
-    }
-
-    public Cliente getCliente() {
-        return cliente;
-    }
-
-    public void setCliente(Cliente cliente) {
-        this.cliente = cliente;
-    }
-
-    public Entregador getEntregador() {
-        return entregador;
-    }
-
-    public void setEntregador(Entregador entregador) {
-        this.entregador = entregador;
-    }
-
-    public Restaurante getRestaurante() {
-        return restaurante;
-    }
-
-    public void setRestaurante(Restaurante restaurante) {
-        this.restaurante = restaurante;
-    }
-
-    public ArrayList<ItemPedido> getItensPedido() {
-        return itensPedido;
-    }
-
-    public void setItensPedido(ArrayList<ItemPedido> itensPedido) {
-        this.itensPedido = itensPedido;
-    }
-
-    public Date getData() {
-        return data;
-    }
-
-    public void setData(Date data) {
-        this.data = data;
-    }
-
-    public BigDecimal getValorTotal() {
-        return valorTotal;
-    }
-
-    public void setValorTotal(BigDecimal valorTotal) {
-        this.valorTotal = valorTotal;
-    }
-
-    public StatusPedido getStatusPedido() {
-        return statusPedido;
-    }
-
-    public void setStatusPedido(StatusPedido statusPedido) {
-        this.statusPedido = statusPedido;
-    }
-}
