--- app/src/main/java/dominio/pedido/StatusPedido.java
@@ -1,10 +1,0 @@
-package dominio.pedido;
-
-public class StatusPedido {
-    public static final int EMESPERA = 1;
-    public static final int EMPREPARO = 2;
-    public static final int SAIUENTREGA = 3;
-    public static final int ENTREGUE = 4;
-    public static final int RECUSADO = 5;
-
-}
