--- app/src/main/java/dominio/restaurante/Ingrediente.java
@@ -1,31 +1,0 @@
-package dominio.restaurante;
-
-public class Ingrediente {
-    private long idIngrediente;
-    private String nome;
-    private boolean disponivel;
-
-    public long getIdIngrediente() {
-        return idIngrediente;
-    }
-
-    public void setIdIngrediente(long idIngrediente) {
-        this.idIngrediente = idIngrediente;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public boolean isDisponivel() {
-        return disponivel;
-    }
-
-    public void setDisponivel(boolean disponivel) {
-        this.disponivel = disponivel;
-    }
-}
